requirements.txt
aniso8601==3.0.2
bcrypt==3.1.4
bson==0.5.6
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.5
cycler==0.10.0
dask==0.19.1
decorator==4.3.0
dnspython==1.15.0
eventlet==0.24.1
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Login==0.4.1
Flask-PyMongo==2.1.0
Flask-RESTful==0.3.6
Flask-Session==0.3.1
Flask-SocketIO==3.0.2
greenlet==0.4.15
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.3
monotonic==1.5
networkx==2.1
numpy==1.15.1
opencv-python==3.4.3.18
Pillow==5.2.0
pycparser==2.18
pymongo==3.7.1
pyparsing==2.2.0
python-dateutil==2.7.3
python-engineio==2.2.0
python-socketio==2.0.0
pytz==2018.5
PyWavelets==1.0.0
recordclass==0.5
requests==2.19.1
scikit-image==0.14.0
scipy==1.1.0
six==1.11.0
toolz==0.9.0
urllib3==1.23
Werkzeug==0.14.1
